Makefile: run test after install
This way to perform tests is flawed in so many ways it becomes weird.

First, building in place do not works for an obscure reason i can't
understood, because copying files from build to gerobust/ source directory
failed with:

    error: could not create 'gerobust/_wrapped_geolib/py.abi3.so': No such file or directory

Therefore, no testing can be done in place because of the absence of
module itself. Installation is necessary.

Secondly, running pytest on gerobust itself leads to an ImportError, as
the wrapped_geolib.py module is not available (remember: the compilation
failed).
Therefore, tests must absolutely be put outside of the package, which is
minor inconvenience.
The pitfall is that doctests in sources CANNOT be executed by pytest.

This is a *very* unpleasant downside.